+++ server/schemas/typeDefs.js
@@ -29,7 +29,7 @@ const typeDefs = gql`
     orders: [Order]
     sales: [Order]
     purchases: [Order]
-    seeds: Int
+    seeds: Float
     products: [Product]
     location: String
   }
@@ -58,10 +58,10 @@ const typeDefs = gql`
     checkout(products: [ID]!): Checkout 
   }
   type Mutation {
-    addUser(firstName: String!, lastName: String!, email: String!, password: String!, location: String!, seeds: Int!): Auth
+    addUser(firstName: String!, lastName: String!, email: String!, password: String!, location: String!, seeds: Float!): Auth
     addOrder(products: [ID]!): Order
     addProduct(name: String!, description: String!, price: Float!, quantity: Int!, category: String!): User
-    addSeeds(_id: ID!, seeds: Int!): User
+    addSeeds(_id: ID!, seeds: Float!): User
     updateUser(firstName: String, lastName: String, email: String, password: String): User
     updateProduct(_id: ID!, quantity: Int!): Product
     login(email: String!, password: String!): Auth
